Added code for update and delete

+++ src/main/java/com/digital/bookstore/service/BookService.java
@@ -2,6 +2,7 @@ package com.digital.bookstore.service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +12,10 @@ import org.springframework.transaction.annotation.Transactional;
 import com.digital.bookstore.entity.Book;
 import com.digital.bookstore.repository.BookRepository;
 import com.digital.bookstore.request.BookRequest;
+import com.digital.bookstore.request.Checkout;
+import com.digital.bookstore.request.CheckoutRequest;
 import com.digital.bookstore.response.BookResponse;
+import com.digital.bookstore.response.CheckoutResponse;
 import com.digital.bookstore.response.Status;
 import static com.digital.bookstore.helper.BookHelper.BOOK_HELPER;
 
@@ -34,23 +38,55 @@ public class BookService {
 	}*/
 
 	 @Transactional
-	public Optional<Status> add(final BookRequest BookRequest) {
+	public Optional<Status> add(final BookRequest bookRequest) {
 		boolean status = true;
 		String message = "Successfully added the book record";
-		Optional<Book> book = BOOK_HELPER.constructBook(BookRequest);
+		Optional<Book> book = BOOK_HELPER.constructBook(bookRequest);
 		book.ifPresent(tmpbook ->bookRepository.save(tmpbook));
 		return BOOK_HELPER.constructStatus(status,message);
 	}
 
 	 
-	/*public Optional<Status> delete(int id) {
+	public Optional<Status> delete(final Integer id) {
+		boolean status = false;
+		String message = "Unable to delete.";
+		if(Objects.isNull(id)) {
+			return BOOK_HELPER.constructStatus(status,message);
+		}
 		bookRepository.deleteById(id);
+		status =  true;
+		message = "Sucessfully deleted";
+		return BOOK_HELPER.constructStatus(status,message);
 	}
 
 	 
-	public Optional<Status> update(BookRequest Book) {
-		boolean status = true;
-		String message = "Successfully update the book record";
-		bookRepository.save(Book);
-	}*/
+	public Optional<Status> update(final BookRequest bookRequest) {
+		
+		
+		boolean status = false;
+		String message = "Unable to update either request is null or book id is null";
+		if(Objects.isNull(bookRequest) || bookRequest.getId() == null || bookRequest.getId() == 0) {
+			return BOOK_HELPER.constructStatus(status,message);
+		}
+		
+		Optional<Book> bookexist = bookRepository.findById(bookRequest.getId());
+		if(!bookexist.isPresent()) {
+			message = "Unable to update. Book does not exist";
+			return BOOK_HELPER.constructStatus(status,message);
+		}
+		Optional<Book> book = BOOK_HELPER.constructBook(bookRequest);
+		book.ifPresent(tmpbook ->bookRepository.save(tmpbook));
+		status =  true;
+		message = "Sucessfully updated";
+		return BOOK_HELPER.constructStatus(status,message);
+	}
+	 
+	 public Optional<CheckoutResponse> checkout(final CheckoutRequest checkoutRequest) {
+		 List<Checkout> checkout = checkoutRequest.getCheckout();
+		 Objects.nonNull(checkout);
+		// checkout.stream().
+		 
+		 return null; 
+		 
+	 }
 }
